discord_bot/requirements.txt
discord.py>=2.3.0
aiohttp>=3.8.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
asyncio-mqtt>=0.16.0

# Local package
-e ../

# Optional dependencies for enhanced features
# google-cloud-language>=2.9.0  # For Perspective API alternative
# redis>=4.5.0  # For distributed caching
# prometheus-client>=0.16.0  # For metrics export